+++ NEWS.md
@@ -1,3 +1,14 @@
+## mmconvert 0.2-3 (2021-10-12)
+
+- Added a dataset with the MUGA array annotations for markers on the
+  autosomes or X chromosome, with mouse build GRCm39 positions and
+  the revised Cox Map genetic map locations.
+
+- Add function `cross2_to_grcm39()` for converting an R/qtl2 cross2
+  object to use the new GRCm39 mouse build and the revised Cox genetic
+  map.
+
+
 ## mmconvert 0.1-5 (2021-09-28)
 
 - New package